% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.r
\name{add_SnaptronMeta}
\alias{add_SnaptronMeta}
\title{Add metadata to file with samples}
\usage{
add_SnaptronMeta(df, compilation)
}
\arguments{
\item{df}{data.frame that includes Snaptron sampleIDs - column should be named "sampleID". Please subset this in some fashion - otherwise it will take forever to download.}

\item{compilation}{Compilation of interest (sra_human, sra_mouse, tcga, gtex, encode).}
}
\value{
A dataframe with five columns ("chromosome", "start", "end", "strand", "samples")
}
\description{
This function downloads metadata for samples of interest.
}
\references{
Christopher Wilks, Phani Gaddipati, Abhinav Nellore, Ben Langmead.
Snaptron: querying splicing patterns across tens of thousands of RNA-seq samples.
Bioinformatics 34, 114-116 (2018). https://doi.org/10.1093/bioinformatics/btx547
}
