% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_psi.R
\name{calcPSI_bulk}
\alias{calcPSI_bulk}
\title{Calculate PSI values for each junction of interest from an info file}
\usage{
calcPSI_bulk(info_df, flatten_wide = F, flatten_long = F)
}
\arguments{
\item{info_df}{Path to the info file which contains (at least) the following columns: novel_junc_id, compilation (sra_human, sra_mouse, tcga, gtex, encode), strand, novel_junc_left_coord, novel_junc_right_coord, canon_junc_coord.}

\item{flatten_wide}{Use flatten_counts function to create wide df. Default: F.}

\item{flatten_long}{Use flatten_counts function to create long df. Default: F.}
}
\value{
List of dataframes - one per each unique target in info_df.
}
\description{
This function calculates PSI values for cryptic exons given an info file
}
\references{
Christopher Wilks, Phani Gaddipati, Abhinav Nellore, Ben Langmead.
Snaptron: querying splicing patterns across tens of thousands of RNA-seq samples.
Bioinformatics 34, 114-116 (2018). https://doi.org/10.1093/bioinformatics/btx547

Sinha, I.R., Sandal, P.S., Spence, H. et al.
Large-scale RNA-Seq mining reveals ciclopirox olamine induces TDP-43 cryptic exons.
Nat Commun 16, 6878 (2025). https://doi.org/10.1038/s41467-025-62004-5
}
