% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_psi.R
\name{calculatePSI}
\alias{calculatePSI}
\title{Calculate PSI values for a single junction of interest}
\usage{
calculatePSI(
  novel_junc_coord,
  canon_junc_coord,
  snaptron_df,
  totalCountMin = 15,
  filter_noCanon = T
)
}
\arguments{
\item{novel_junc_coord}{Coordinates of the novel junction of interest (chr\link{1-22XYM}:\link{start}-\link{end})}

\item{canon_junc_coord}{Coordinates of the related canonical junction (chr\link{1-22XYM}:\link{start}-\link{end})}

\item{snaptron_df}{Output from snaptronQuery function.}

\item{totalCountMin}{Sum of reads of canonical and novel junctions. Default: 15.}

\item{filter_noCanon}{Filter out samples with 0 reads of canonical junction. Default: T.}
}
\value{
PSI value of given junction of interest within all Snaptron samples
}
\description{
This function calculates PSI values for cryptic exons given junction coordinates
}
\references{
Christopher Wilks, Phani Gaddipati, Abhinav Nellore, Ben Langmead.
Snaptron: querying splicing patterns across tens of thousands of RNA-seq samples.
Bioinformatics 34, 114-116 (2018). https://doi.org/10.1093/bioinformatics/btx547

Sinha, I.R., Sandal, P.S., Spence, H. et al.
Large-scale RNA-Seq mining reveals ciclopirox olamine induces TDP-43 cryptic exons.
Nat Commun 16, 6878 (2025). https://doi.org/10.1038/s41467-025-62004-5
}
