% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_counts.R
\name{flattenCounts}
\alias{flattenCounts}
\title{Take list of count dataframes and turn it into a single dataframe}
\usage{
flattenCounts(counts_list, wide = T)
}
\arguments{
\item{counts_list}{Path to the info file which contains (at least) the following columns: id, compilation (sra_human, sra_mouse, tcga, gtex, encode), strand, novel_junc_left_coord, novel_junc_right_coord, canon_junc_coord.}

\item{wide}{T/F value - whether dataframe should have one column per novel junction or be a long df compatible with packages such as ggplot2}
}
\value{
Single dataframe
}
\description{
This function merges the counts dfs generated by calcPSI_bulk
}
\references{
Christopher Wilks, Phani Gaddipati, Abhinav Nellore, Ben Langmead.
Snaptron: querying splicing patterns across tens of thousands of RNA-seq samples.
Bioinformatics 34, 114-116 (2018). https://doi.org/10.1093/bioinformatics/btx547

Sinha, I.R., Sandal, P.S., Spence, H. et al.
Large-scale RNA-Seq mining reveals ciclopirox olamine induces TDP-43 cryptic exons.
Nat Commun 16, 6878 (2025). https://doi.org/10.1038/s41467-025-62004-5
}
