% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_counts.R
\name{junctionData}
\alias{junctionData}
\title{Download Snaptron junction counts based on given junction, strand, and compilation}
\usage{
junctionData(junction, strand, compilation)
}
\arguments{
\item{junction}{Junction coordinates to search Snaptron database. Usually "canon_junc_coord" from info file.}

\item{strand}{Chromosome strand on which the gene of interest is located (+ or -). Usually "strand" from info file.}

\item{compilation}{Compilation of interest (sra_human, sra_mouse, tcga, gtex, encode),. Usually "compilation" from info file.}
}
\value{
A dataframe with five columns ("chromosome", "start", "end", "strand", "samples")
}
\description{
This function creates the Snaptron url for a given novel_junc_id
}
\references{
Christopher Wilks, Phani Gaddipati, Abhinav Nellore, Ben Langmead.
Snaptron: querying splicing patterns across tens of thousands of RNA-seq samples.
Bioinformatics 34, 114-116 (2018). https://doi.org/10.1093/bioinformatics/btx547

Sinha, I.R., Sandal, P.S., Spence, H. et al.
Large-scale RNA-Seq mining reveals ciclopirox olamine induces TDP-43 cryptic exons.
Nat Commun 16, 6878 (2025). https://doi.org/10.1038/s41467-025-62004-5
}
