% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_counts.R
\name{sampleJunctionCounts}
\alias{sampleJunctionCounts}
\title{Extract Junction Counts for each Sample}
\usage{
sampleJunctionCounts(df)
}
\arguments{
\item{df}{Path to the input file which is a dataframe from Snaptron.}
}
\value{
A dataframe with two columns (rail_ids, counts).
}
\description{
This function extracts junction count information from Snaptron-derived dataframe. It creates a new dataframe with 1 column of rail_ids and 1 column of counts.
}
\references{
Christopher Wilks, Phani Gaddipati, Abhinav Nellore, Ben Langmead.
Snaptron: querying splicing patterns across tens of thousands of RNA-seq samples.
Bioinformatics 34, 114-116 (2018). https://doi.org/10.1093/bioinformatics/btx547

Sinha, I.R., Sandal, P.S., Spence, H. et al.
Large-scale RNA-Seq mining reveals ciclopirox olamine induces TDP-43 cryptic exons.
Nat Commun 16, 6878 (2025). https://doi.org/10.1038/s41467-025-62004-5
}
