% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_counts.R
\name{snaptronQuery}
\alias{snaptronQuery}
\title{Download junction data from Snaptron for all queried junctions}
\usage{
snaptronQuery(uniqueCanonJunc)
}
\arguments{
\item{uniqueCanonJunc}{Path to the dataframe containing four columns: junction (one row per each unique canonical junction), strand, compilation, and allJunc (all junctions queried within that canonical junction)}
}
\value{
A dataframe with three columns (junction, strand, samples)
}
\description{
This function extracts junction count information from Snaptron.
}
\references{
Christopher Wilks, Phani Gaddipati, Abhinav Nellore, Ben Langmead.
Snaptron: querying splicing patterns across tens of thousands of RNA-seq samples.
Bioinformatics 34, 114-116 (2018). https://doi.org/10.1093/bioinformatics/btx547

Sinha, I.R., Sandal, P.S., Spence, H. et al.
Large-scale RNA-Seq mining reveals ciclopirox olamine induces TDP-43 cryptic exons.
Nat Commun 16, 6878 (2025). https://doi.org/10.1038/s41467-025-62004-5
}
